Created master CONTRIBUTING.md file
This file is a combination of both the original CONTRIBUTING.md
file from Codidact/Core and the Commit Message Guidelines document.
The reason for the combination is the feeling that one of the two
documents always goes unheeded (I am guilty of this myself) due to
people not realizing there are two documents describing the
contribution process. Merging these two documents should alleviate
the number of reprimands Marc needs to make on pull requests c:

This commit fixes issue #3