Added README section describing the purpose of this repository
The README was initially modified to include a Table of Contents
and a section for each of the three critical files the repository
will initially contain. It did not, however, do a good job of
explaining exactly what the purpose was for having a separate
repository for holding files which would normally be placed in a
project directory. Hopefully this commit explains the necessity
for this repository in a clear and concise manner.

This commit fixes issue #4