Table of Contents and Headers are now links
The table of contents entries may now be used to navigate the
README itself, and the headers of each section may be used to
navigate to the actual community file in question.

In addition, I added a section at the end of the README directing
all questions to the Codidact Team Leads, and all hate mail and/or
corrections to me.

This commit fixes issue #8